Goodwill guesture
Probably a good idea to change out the Twitter handle if you're going to snag our script. :) It may also be a good idea to go private with this repository; we try to turn our heads a bit when people use our work, but we would prefer that it doesn't get open-sourced.

All of that said, what you've done with this site is certainly a well executed re-skinning. Keep up the good work!